refact: notification service 리팩토링
